+++ CHANGELOG.md
@@ -37,6 +37,7 @@ Additions
 - Allow specifying a maximum amount of characters for a TextField
 - Added a multiline editing mode to TextField
 - Added a formatting code to allow for inline font changes
+- Added a SquishingGroup element
 
 Improvements
 - **Made Image ScaleToImage take ui scale into account**